Extract all the network methods into a vmware_network utils file
