+++ plugins/modules/vm_portgroup_info.py
@@ -9,6 +9,7 @@ from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
 
+
 DOCUMENTATION = r'''
 ---
 module: vm_portgroup_info
@@ -65,98 +66,13 @@ vm_portgroup_info:
 
 from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.vmware.vmware.plugins.module_utils.vmware import PyVmomi
+from ansible_collections.vmware.vmware.plugins.module_utils import vmware_network
 from ansible_collections.vmware.vmware.plugins.module_utils.vmware_rest_client import VmwareRestClient
-
-try:
-    from pyVmomi import vim
-except ImportError:
-    pass
-
-
-def get_standard_portgroup_vlan_vswitch(portgroup, pg_name):
-    ret_pg = {'name': pg_name}
-    for host in portgroup.host:
-        pgs = host.config.network.portgroup
-        for pg in pgs:
-            if pg.spec.name == pg_name:
-                ret_pg['vlan_id'] = str(pg.spec.vlanId)
-                ret_pg['vswitch_name'] = str(pg.spec.vswitchName)
-                return ret_pg
-
-
-def get_teaming_policy(uplink_teaming_policy):
-    return dict(
-        policy=uplink_teaming_policy.policy.value,
-        inbound_policy=uplink_teaming_policy.reversePolicy.value,
-        notify_switches=uplink_teaming_policy.notifySwitches.value,
-        rolling_order=uplink_teaming_policy.rollingOrder.value,
-    )
-
-
-def get_port_policy(config_policy):
-    return dict(
-        block_override=config_policy.blockOverrideAllowed,
-        ipfix_override=config_policy.ipfixOverrideAllowed,
-        live_port_move=config_policy.livePortMovingAllowed,
-        network_rp_override=config_policy.networkResourcePoolOverrideAllowed,
-        port_config_reset_at_disconnect=config_policy.portConfigResetAtDisconnect,
-        security_override=config_policy.macManagementOverrideAllowed,
-        shaping_override=config_policy.shapingOverrideAllowed,
-        traffic_filter_override=config_policy.trafficFilterOverrideAllowed,
-        uplink_teaming_override=config_policy.uplinkTeamingOverrideAllowed,
-        vendor_config_override=config_policy.vendorConfigOverrideAllowed,
-        vlan_override=config_policy.vlanOverrideAllowed
-    )
-
-
-def get_dvs_mac_learning(mac_learning_policy):
-    return dict(
-        allow_unicast_flooding=mac_learning_policy.allowUnicastFlooding,
-        enabled=mac_learning_policy.enabled,
-        limit=mac_learning_policy.limit,
-        limit_policy=mac_learning_policy.limitPolicy
-    )
-
-
-def get_dvs_network_policy(mac_management_policy):
-    return dict(
-        forged_transmits=mac_management_policy.forgedTransmits,
-        promiscuous=mac_management_policy.allowPromiscuous,
-        mac_changes=mac_management_policy.macChanges
-    )
-
-
-def get_vlan_info(vlan_obj):
-    if isinstance(vlan_obj, vim.dvs.VmwareDistributedVirtualSwitch.TrunkVlanSpec):
-        vlan_id_list = []
-        for vli in vlan_obj.vlanId:
-            if vli.start == vli.end:
-                vlan_id_list.append(str(vli.start))
-            else:
-                vlan_id_list.append('{}-{}'.format(vli.start, vli.end))
-        return dict(trunk=True, pvlan=False, vlan_id=vlan_id_list)
-    elif isinstance(vlan_obj, vim.dvs.VmwareDistributedVirtualSwitch.PvlanSpec):
-        return dict(trunk=False, pvlan=True, vlan_id=str(vlan_obj.pvlanId))
-    else:
-        return dict(trunk=False, pvlan=False, vlan_id=str(vlan_obj.vlanId))
-
-
-def get_dvs_port_allocation(config_type):
-    if config_type == 'ephemeral':
-        return 'ephemeral'
-    else:
-        return 'static'
-
-
-def get_dvs_autoExpand(config_autoexpand):
-    return 'elastic' if config_autoexpand else 'fixed'
 
 
 class PortgroupInfo(PyVmomi):
     def __init__(self, module):
         super(PortgroupInfo, self).__init__(module)
-        self.module = module
-        self.params = module.params
         self.vmware_client = VmwareRestClient(module)
         self.vms = self.params['vm_names']
 
@@ -164,13 +80,13 @@ class PortgroupInfo(PyVmomi):
         dvs_pg = self.get_dvs_portgroup(pg_id)
         pg = {'portgroup_name': dvs_pg.name, 'vswitch_name': dvs_pg.config.distributedVirtualSwitch.name,
               'type': 'DISTRIBUTED_PORTGROUP', 'port_id': pg_id,
-              'port_binding': get_dvs_port_allocation(dvs_pg.config.type),
-              'port_allocation': get_dvs_autoExpand(dvs_pg.config.autoExpand),
-              'network_policy': get_dvs_network_policy(dvs_pg.config.defaultPortConfig.macManagementPolicy),
-              'mac_learning': get_dvs_mac_learning(dvs_pg.config.defaultPortConfig.macManagementPolicy.macLearningPolicy),
-              'teaming_policy': get_teaming_policy(dvs_pg.config.defaultPortConfig.uplinkTeamingPolicy),
-              'port_policy': get_port_policy(dvs_pg.config.policy),
-              'vlan_info': get_vlan_info(dvs_pg.config.defaultPortConfig.vlan)
+              'port_binding': vmware_network.get_dvs_port_allocation(dvs_pg.config.type),
+              'port_allocation': vmware_network.get_dvs_auto_expand(dvs_pg.config.autoExpand),
+              'network_policy': vmware_network.get_dvs_network_policy(dvs_pg.config.defaultPortConfig.macManagementPolicy),
+              'mac_learning': vmware_network.get_dvs_mac_learning(dvs_pg.config.defaultPortConfig.macManagementPolicy.macLearningPolicy),
+              'teaming_policy': vmware_network.get_teaming_policy(dvs_pg.config.defaultPortConfig.uplinkTeamingPolicy),
+              'port_policy': vmware_network.get_port_policy(dvs_pg.config.policy),
+              'vlan_info': vmware_network.get_vlan_info(dvs_pg.config.defaultPortConfig.vlan)
               }
 
         if dvs_pg.config.defaultPortConfig.uplinkTeamingPolicy and \
@@ -183,7 +99,7 @@ class PortgroupInfo(PyVmomi):
     def get_standard_portgroup_detailed(self, pg_id):
         pg = self.get_standard_portgroup(pg_id)
         pg_name = str(pg.summary.name)
-        ret_pg = get_standard_portgroup_vlan_vswitch(pg, pg_name)
+        ret_pg = vmware_network.get_standard_portgroup_vlan_vswitch(pg, pg_name)
         ret_pg['port_id'] = pg_id
         ret_pg['type'] = 'STANDARD_PORTGROUP'
         return ret_pg
